# openapi.yaml
openapi: 3.1.0
info:
  title: GPT Memory API
  description: |
    REST API for persistent memory management of tabletop RPG campaigns.
    
    Supports thousands of independent campaigns. All entities are scoped to a campaign via `campaignId`.
    All responses are JSON. Soft-delete is enforced across all entities.
  version: 1.0.0
  contact:
    name: GPT Memory

servers:
  - url:   https://gpt-memory.jokertcnr.workers.dev/api
    description: Production
  - url: http://localhost:8787/api
    description: Local development

tags:
  - name: Health
  - name: Campaigns
  - name: Characters
  - name: Locations
  - name: Planets
  - name: StarSystems
  - name: Organizations
  - name: Ships
  - name: Vehicles
  - name: Objects
  - name: Events
  - name: Timeline
  - name: Relationships
  - name: Secrets
  - name: PlayerKnowledge
  - name: WorldKnowledge
  - name: Quotes
  - name: Notes
  - name: Sessions
  - name: Search
  - name: Context

components:
  parameters:
    campaignId:
      in: query
      name: campaignId
      required: true
      schema:
        type: string
      description: The campaign UUID to scope the query to
    id:
      in: path
      name: id
      required: true
      schema:
        type: string
    page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
        metadata:
          type: object
          nullable: true

    Character:
      type: object
      properties:
        id:
          type: string
        campaignId:
          type: string
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
          nullable: true
        species:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        birth:
          type: string
          nullable: true
        death:
          type: string
          nullable: true
        status:
          type: string
          enum: [alive, dead, unknown, missing]
        organization:
          type: string
          nullable: true
        rank:
          type: string
          nullable: true
        currentPlanet:
          type: string
          nullable: true
        currentLocation:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        emotionalState:
          type: string
          nullable: true
        trust:
          type: number
        respect:
          type: number
        friendship:
          type: number
        fear:
          type: number
        curiosity:
          type: number
        anger:
          type: number
        love:
          type: number
        hate:
          type: number
        debt:
          type: number
        lastInteraction:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

paths:
  /health:
    get:
      operationId: healthCheck
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: Service is healthy

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [Campaigns]
      summary: List all campaigns
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: q
          schema:
            type: string
          description: Search by name
      responses:
        '200':
          description: List of campaigns with pagination
    post:
      operationId: createCampaign
      tags: [Campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Campaign created

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [Campaigns]
      summary: Get a campaign by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign found
        '404':
          description: Campaign not found
    patch:
      operationId: updateCampaign
      tags: [Campaigns]
      summary: Update a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Campaign updated
    delete:
      operationId: deleteCampaign
      tags: [Campaigns]
      summary: Soft-delete a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign deleted

  /characters:
    get:
      operationId: listCharacters
      tags: [Characters]
      summary: List characters in a campaign
      parameters:
        - in: query
          name: campaignId
          required: true
          schema:
            type: string
          description: The campaign UUID to scope the query to
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [alive, dead, unknown, missing]
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of characters
    post:
      operationId: createCharacter
      tags: [Characters]
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Character'
      responses:
        '201':
          description: Character created

  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [Characters]
      summary: Get a character
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: campaignId
          required: true
          schema:
            type: string
          description: The campaign UUID to scope the query to
      responses:
        '200':
          description: Character found
    patch:
      operationId: updateCharacter
      tags: [Characters]
      summary: Update a character
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Character updated
    delete:
      operationId: deleteCharacter
      tags: [Characters]
      summary: Soft-delete a character
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: campaignId
          required: true
          schema:
            type: string
          description: The campaign UUID to scope the query to
      responses:
        '200':
          description: Character deleted

  /search:
    get:
      operationId: search
      tags: [Search]
      summary: Full-text search across all entity types
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Search query (FTS5)
        - in: query
          name: campaignId
          schema:
            type: string
          description: Limit to a specific campaign
        - in: query
          name: types
          schema:
            type: string
          description: Comma-separated entity types to filter
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
        - in: query
          name: tags
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results

  /context:
    get:
      operationId: buildContext
      tags: [Context]
      summary: Build scene context for a given location/time/characters
      parameters:
        - in: query
          name: campaignId
          required: true
          schema:
            type: string
          description: The campaign UUID to scope the query to
        - in: query
          name: planet
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
        - in: query
          name: characters
          schema:
            type: string
          description: Comma-separated character IDs
        - in: query
          name: radius
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Scene context with characters, events, relationships, secrets, and knowledge